* master:
  Retry connecting to mongodb on startup failures.
  Whoops, remove some debug stuff from f3bc4c4ef2f139b2ecdc237ded2a34e61b940101
  Fix some race conditions in tests that led to periodic failures.
  Improve error handling during tests if redis or mongo connections fail.
  Flip order of test matchers to properly reflect "expected" output.
  Standardize on CircleCI for testing.
  Let the import_nginx_logs script uncompress gzipped logs on the fly.
  Revamp logging to use Bunyan and cleanup how Rollbar errors are logged.
  Experimenting with sending more detailed error logs to rollbar.
  Begin refactoring log collection process.

+++ test/support/database_setup.js
@@ -15,15 +15,17 @@ var config = apiUmbrellaConfig.load(path.resolve(__dirname, '../config/test.yml'
 
 mongoose.testConnection = mongoose.createConnection(config.get('mongodb.url'), config.get('mongodb.options'));
 
-// Drop the mongodb database.
 before(function mongoOpen(done) {
-  mongoose.testConnection.on('connected', function() {
-    // Drop the whole database, since that properly blocks for any active
-    // connections. The database will get re-created on demand.
-    mongoose.testConnection.db.dropDatabase(function() {
-      done();
-    });
-  });
+  mongoose.testConnection.on('connected', done);
+});
+
+before(function mongoDropDatabase(done) {
+  this.timeout(5000);
+
+  // Drop the whole database, since that properly blocks for any active
+  // connections. The database will get re-created on demand. If this times
+  // out, it's likely because there are active connections.
+  mongoose.testConnection.db.dropDatabase(done);
 });
 
 // Spin up a new redis-server for running the test suite. Multiple database on
@@ -116,16 +118,18 @@ after(function mongoClose(done) {
 });
 
 after(function redisClose(done) {
-  global.redisClient.quit(function() {
-    if(redisServer.running) {
-      redisServer.on('exit', function() {
-        done();
-      });
+  if(global.redisClient) {
+    global.redisClient.quit(function() {
+      if(redisServer.running) {
+        redisServer.on('exit', function() {
+          done();
+        });
 
-      redisServer.stop();
-      fs.unlinkSync(redisPidFile);
-    } else {
-      done();
-    }
-  });
+        redisServer.stop();
+        fs.unlinkSync(redisPidFile);
+      } else {
+        done();
+      }
+    });
+  }
 });
